+++ changelog.rst
@@ -1,11 +1,12 @@
 Changelog
 =========
 
-Last change: 13-JUN-2025 MTS
+Last change: 15-JUL-2025 MTS
 
 0.8.1
 -----
 - Added ``n_events`` to cluster centers, i.e., number of binding events per cluster
+- .yaml files contain Picasso version number for easier tracking
 
 0.8.0
 -----