+++ changelog.rst
@@ -7,7 +7,8 @@ Last change: 15-JUL-2025 MTS
 -----
 - Added ``n_events`` to cluster centers, i.e., number of binding events per cluster
 - .yaml files contain Picasso version number for easier tracking
-- Bug fixes
+- Improved fiducial picking
+- Bug fixes and other cosmetic changes
 
 0.8.0
 -----